refactor: 오류 수정
