+++ Playproof/src/features/community/hooks/useCommunityPageLogic.ts
@@ -1,195 +1,85 @@
+// @refresh reset
 // src/features/community/hooks/useCommunityPageLogic.ts
 
-import React from "react";
-import { useNavigate, useSearchParams } from "react-router-dom";
-import { useAuthStore } from "@/store/authStore";
-import { COMMUNITY_PAGE_LABELS } from "@/features/community/constants/labels";
-import type { CommunityTab, HighlightPost, BoardPost } from "@/features/community/types";
-import { useHighlightFeed } from "@/features/community/hooks/useHighlightFeed";
-import { useCommunityDataLoad } from "@/features/community/hooks/useCommunityDataLoad";
-import { useCommunityFilters } from "@/features/community/hooks/useCommunityFilters";
-import { useCommunityWrite } from "@/features/community/hooks/useCommunityWrite";
+import { useCommunityUI } from "@/features/community/hooks/useCommunityUI";
+import { useCommunityData } from "@/features/community/hooks/useCommunityData";
 
 export const useCommunityPageLogic = () => {
-  const navigate = useNavigate();
-  const [searchParams, setSearchParams] = useSearchParams();
-  const authUserId = useAuthStore((s) => s.userId);
-  const userId = authUserId ? `user-${authUserId}` : undefined;
-  const {
-    state: highlightState,
-    actions: highlightActions,
-  } = useHighlightFeed();
-
-  const [currentPage, setCurrentPage] = React.useState(1);
-  const [selectedPost, setSelectedPost] = React.useState<HighlightPost | null>(null);
-  const [isModalOpen, setIsModalOpen] = React.useState(false);
-  const [boardGame, setBoardGame] = React.useState("전체글");
-
-  const currentTab = searchParams.get("tab");
-  const activeTab: CommunityTab =
-    currentTab === COMMUNITY_PAGE_LABELS.freeTab
-      ? COMMUNITY_PAGE_LABELS.freeTab
-      : COMMUNITY_PAGE_LABELS.highlightTab;
-  const handleTabChange = (tab: CommunityTab) => {
-    setSearchParams({ tab });
-    setCurrentPage(1);
-  };
-
-  const handleHighlightClick = (post: HighlightPost) => {
-    setSelectedPost(post);
-    setIsModalOpen(true);
-    highlightActions.fetchComments(post.id);
-  };
-
-  const handleBoardClick = (post: BoardPost) => {
-    navigate(`/community/${post.id}?from=자유게시판`, { state: { post } });
-  };
-
-  const handleCloseModal = () => {
-    setIsModalOpen(false);
-    setSelectedPost(null);
-  };
-
-  const BOARD_GAME_ID_MAP: Record<string, number> = {
-    "전체글": 0,
-    "리그오브레전드": 1,
-    "발로란트": 2,
-    "오버워치": 3,
-  };
-  const boardGameId = BOARD_GAME_ID_MAP[boardGame] ?? 0;
-
-  const {
-    boardPosts: loadedBoardPosts,
-    setBoardPosts: setLoadedBoardPosts,
-    bestPosts: loadedBestPosts,
-    loading: loadedLoading,
-  } = useCommunityDataLoad({
-    activeTab,
-    currentPage,
-    boardGameId,
-    hydrateHighlights: highlightActions.hydrateFromPosts,
+  const { ui: uiState, setters, actions: uiActions } = useCommunityUI();
+  const dataState = useCommunityData({
+    activeTab: uiState.activeTab,
+    currentPage: uiState.currentPage,
+    boardGame: uiState.boardGame,
+    setBoardGame: setters.setBoardGame,
+    setSelectedPost: setters.setSelectedPost,
+    setIsModalOpen: setters.setIsModalOpen,
   });
-
-  const {
-    searchQuery,
-    setSearchQuery,
-    handleSearch,
-    isFilterOpen,
-    setIsFilterOpen,
-    filters,
-    setFilters,
-    filteredHighlights,
-    filteredBoardPosts,
-    totalPages,
-  } = useCommunityFilters({
-    highlights: highlightState.highlights,
-    boardPosts: loadedBoardPosts,
-    itemsPerPage: 10,
-    boardGame,
-    setBoardGame,
-  });
-
-  const {
-    isWriteOpen,
-    setIsWriteOpen,
-    handleWritePost,
-    handleWriteSubmit,
-    revokeBoardMedia,
-  } = useCommunityWrite({
-    activeTab,
-    boardGame,
-    refreshHighlights: () => {
-      // 하이라이트 생성 후 목록 새로고침
-      import("@/features/community/api/communityApi").then(({ getHighlights }) => {
-        getHighlights(currentPage).then((posts) => {
-          highlightActions.hydrateFromPosts(posts);
-        });
-      });
-    },
-    refreshBoardPosts: () => {
-      // 자유게시판 글 생성 후 목록 새로고침
-      import("@/features/community/api/communityApi").then(({ getBoardPosts, getAllBoardPosts }) => {
-        const fetchPosts = boardGameId === 0
-          ? getAllBoardPosts(currentPage)
-          : getBoardPosts(boardGameId, currentPage);
-        fetchPosts.then((posts) => {
-          setLoadedBoardPosts(posts);
-        });
-      });
-    },
-  });
-
-  React.useEffect(() => {
-    return () => {
-      revokeBoardMedia();
-    };
-  }, [revokeBoardMedia]);
 
   return {
     ui: {
-      activeTab,
-      searchQuery,
-      currentPage,
-      isFilterOpen,
-      filters,
-      boardGame,
+      activeTab: uiState.activeTab,
+      searchQuery: dataState.ui.searchQuery,
+      currentPage: uiState.currentPage,
+      isFilterOpen: dataState.ui.isFilterOpen,
+      filters: dataState.ui.filters,
+      boardGame: uiState.boardGame,
     },
     data: {
-      bestPosts: loadedBestPosts,
-      loading: loadedLoading,
-      filteredHighlights,
-      filteredBoardPosts,
-      totalPages,
+      bestPosts: dataState.data.bestPosts,
+      loading: dataState.data.loading,
+      filteredHighlights: dataState.data.filteredHighlights,
+      filteredBoardPosts: dataState.data.filteredBoardPosts,
+      totalPages: dataState.data.totalPages,
     },
     modal: {
-      selectedPost,
-      isModalOpen,
-      isWriteOpen,
+      selectedPost: uiState.selectedPost,
+      isModalOpen: uiState.isModalOpen,
+      isWriteOpen: dataState.modal.isWriteOpen,
     },
     user: {
-      currentUserName: highlightState.currentUserName,
-      userId,
+      currentUserName: dataState.user.currentUserName,
+      userId: dataState.user.userId,
     },
     actions: {
       tab: {
-        change: handleTabChange,
+        change: uiActions.tab.change,
       },
       search: {
-        setQuery: setSearchQuery,
-        submit: handleSearch,
+        setQuery: dataState.actions.search.setQuery,
+        submit: dataState.actions.search.submit,
       },
       filter: {
-        setOpen: setIsFilterOpen,
-        setFilters,
+        setOpen: dataState.actions.filter.setOpen,
+        setFilters: dataState.actions.filter.setFilters,
       },
       board: {
-        openDetail: handleBoardClick,
-        setGame: setBoardGame,
+        openDetail: uiActions.board.openDetail,
+        setGame: uiActions.board.setGame,
       },
       pagination: {
-        setPage: setCurrentPage,
+        setPage: uiActions.pagination.setPage,
       },
       highlight: {
-        openDetail: handleHighlightClick,
-        getLikeState: highlightActions.getLikeState,
-        getComments: highlightActions.getComments,
-        getCommentCount: highlightActions.getCommentCount,
-        toggleLike: highlightActions.toggleLike,
-        addComment: highlightActions.addComment,
-        addReply: highlightActions.addReply,
-        editComment: highlightActions.editComment,
-        editReply: highlightActions.editReply,
-        deleteComment: highlightActions.deleteComment,
-        deleteReply: highlightActions.deleteReply,
-        deletePost: highlightActions.deletePost,
+        openDetail: dataState.actions.highlight.openDetail,
+        getLikeState: dataState.actions.highlight.getLikeState,
+        getComments: dataState.actions.highlight.getComments,
+        getCommentCount: dataState.actions.highlight.getCommentCount,
+        toggleLike: dataState.actions.highlight.toggleLike,
+        addComment: dataState.actions.highlight.addComment,
+        addReply: dataState.actions.highlight.addReply,
+        editComment: dataState.actions.highlight.editComment,
+        editReply: dataState.actions.highlight.editReply,
+        deleteComment: dataState.actions.highlight.deleteComment,
+        deleteReply: dataState.actions.highlight.deleteReply,
+        deletePost: dataState.actions.highlight.deletePost,
       },
       modal: {
-        closeDetail: handleCloseModal,
-        setWriteOpen: setIsWriteOpen,
+        closeDetail: uiActions.modal.closeDetail,
+        setWriteOpen: dataState.actions.modal.setWriteOpen,
       },
       write: {
-        open: handleWritePost,
-        submit: handleWriteSubmit,
+        open: dataState.actions.write.open,
+        submit: dataState.actions.write.submit,
       },
     },
   };
